extract Serialized<T> util
