+++ src/vs/base/common/types.ts
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { URI, UriComponents } from 'vs/base/common/uri';
+
 const _typeof = {
 	number: 'number',
 	string: 'string',
@@ -262,3 +264,12 @@ export type AddFirstParameterToFunctions<Target, TargetFunctionsReturnType, Firs
 	// Else: just leave as is
 	Target[K]
 };
+
+/**
+ * Mapped typed that replaces all occurrences of URI with UriComponents and drop all functions
+ */
+export type Serialized<T> = { [K in keyof T]: T[K] extends URI
+	? UriComponents
+	: T[K] extends Function
+	? never
+	: Serialized<T[K]> };
