+++ src/vs/platform/actions/common/actions.ts
@@ -11,8 +11,9 @@ import { ContextKeyExpr, IContextKeyService } from 'vs/platform/contextkey/commo
 import { ICommandService, CommandsRegistry, ICommandHandlerDescription } from 'vs/platform/commands/common/commands';
 import { IDisposable, DisposableStore } from 'vs/base/common/lifecycle';
 import { Event, Emitter } from 'vs/base/common/event';
-import { URI, UriComponents } from 'vs/base/common/uri';
+import { URI } from 'vs/base/common/uri';
 import { ThemeIcon } from 'vs/platform/theme/common/themeService';
+import { Serialized } from 'vs/base/common/types';
 
 export interface ILocalizedString {
 	value: string;
@@ -27,8 +28,6 @@ export interface ICommandAction {
 	precondition?: ContextKeyExpr;
 	toggled?: ContextKeyExpr;
 }
-
-type Serialized<T> = { [K in keyof T]: T[K] extends URI ? UriComponents : Serialized<T[K]> };
 
 export type ISerializableCommandAction = Serialized<ICommandAction>;
